+++ app/jobs/Messages/expiration_job.rb
@@ -1,0 +1,11 @@
+module Messages
+  class ExpirationJob < ApplicationJob
+    queue_as :default
+
+    def perform(message_id)
+      Messages::Expirer.new(message_id).call
+    rescue StandardError => e
+      Rails.logger.error("Message with id #{e} not found")
+    end
+  end
+end
