+++ spec/jobs/messages/expiration_job_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe MessageExpirationJob, type: :job do
+RSpec.describe Messages::ExpirationJob, type: :job do
   after do
     clear_enqueued_jobs
     clear_performed_jobs
