+++ spec/support/shared_examples/successful_time_based_message_creation.rb
@@ -1,6 +1,6 @@
 RSpec.shared_examples 'successful time based message creation' do
   it 'returns true' do
-    expect(creator.call).to be_instance_of(MessageExpirationJob)
+    expect(creator.call).to be_instance_of(Messages::ExpirationJob)
   end
 
   it 'creates a Message instance in message attribute' do
@@ -8,7 +8,7 @@ RSpec.shared_examples 'successful time based message creation' do
     expect(creator.message).to be_instance_of(Message)
   end
 
-  it 'enqueues a MessageExpirationJob' do
-    expect { creator.call }.to have_enqueued_job(MessageExpirationJob)
+  it 'enqueues a Messages::ExpirationJob' do
+    expect { creator.call }.to have_enqueued_job(Messages::ExpirationJob)
   end
 end
