+++ spec/support/shared_examples/successful_visits_based_message_creation.rb
@@ -8,7 +8,7 @@ RSpec.shared_examples 'successful visits based message creation' do
     expect(creator.message).to be_instance_of(Message)
   end
 
-  it 'does not enqueue a MessageExpirationJob' do
-    expect { creator.call }.not_to have_enqueued_job(MessageExpirationJob)
+  it 'does not enqueue a Messages::ExpirationJob' do
+    expect { creator.call }.not_to have_enqueued_job(Messages::ExpirationJob)
   end
 end
